[RMATE-01] - refactor
