+++ app/src/main/java/com/rodcollab/readermate/core/model/BookDto.kt
@@ -1,4 +1,4 @@
-package com.rodcollab.readermate.books
+package com.rodcollab.readermate.core.model
 
 /**
  * Model related to the google books api
